requirements.txt
anthropic
llama-index
llama-index-core
llama-index-llms-anthropic
llama-index-vector-stores-postgres
llama-parse
llama-index-llms-openai 
llama-index-embeddings-openai 
llama-index-embeddings-huggingface
llama-index-readers-file 
llama-index-vector-stores-kdbai
PyPDF2
pandas
jupyter
selenium
beautifulsoup4
webdriver-manager
langdetect
rank_bm25
openai
streamlit
docx2txt
python-docx 
streamlit_authenticator


